requirements.txt
streamlit
pillow
rembg
